pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "narractive"
version = "2.4.0"
description = "Modular framework for automating desktop application demo video production"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Simon Ducorneau"},
]
keywords = ["video", "automation", "pyautogui", "obs", "tts", "ffmpeg", "demo", "screencast", "narractive"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "pyautogui>=0.9.54",
    "Pillow>=10.0",
    "pyyaml>=6.0",
    "click>=8.1",
    "edge-tts>=6.1.0",
    "mutagen>=1.47.0",
]

[project.optional-dependencies]
obs = ["obsws-python>=1.7.0"]
windows = ["pywin32>=306"]
png = ["playwright>=1.40"]
elevenlabs = ["elevenlabs"]
xtts = ["TTS>=0.22"]
kokoro = ["kokoro", "soundfile"]
openai = ["openai>=1.0"]
all = [
    "obsws-python>=1.7.0",
    "pywin32>=306",
    "playwright>=1.40",
    "rich>=13.0",
]
headless = [
    "python-xlib>=0.33",
]
report = ["rich>=13.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
narractive = "narractive.cli:main"

[project.urls]
Homepage = "https://github.com/imagodata/narractive"
Repository = "https://github.com/imagodata/narractive"

[tool.setuptools.packages.find]
include = ["narractive*"]

[tool.setuptools.package-data]
narractive = ["diagrams/*.html"]

# ── Ruff ──────────────────────────────────────────────────────────────────────

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["narractive"]

# ── Mypy ──────────────────────────────────────────────────────────────────────

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# ── Pytest ────────────────────────────────────────────────────────────────────

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
